netgear_router_test/requirements.txt
pytest
pytest-html
paramiko
requests
pysnmp
pyyaml
